tekton/task/cosa-build: initialize
As part of [1], non-core Konflux tasks and pipelines will move out to
external repos (i.e outside of [2]). We're anticipating the switch by
defining our task here instead of [2].
We are taking advantage of Konflux and `tekton-bundle-builder pipeline`
in order to build and push our tekton bundle.
Once the `external-tasks/` scheme ready, we'll push the task bundle
URL in [2] as well for the Konflux community (e.g [3]).

The task was initially developped in [4].

[1] https://issues.redhat.com/browse/KONFLUX-2735
[2] https://github.com/konflux-ci/build-definitions
[3] https://github.com/konflux-ci/build-definitions/pull/2309
[4] https://github.com/jcapiitao/coreos-konflux-definitions/tree/main/tasks/cosa-builder-container