build: allow SKIP_UNTESTED_ARTIFACTS to be null
This allows the parameter to be unset (empty string), which is used
to fall back to the dynamic default determined by
`should_we_skip_untested_artifacts()`.

We cannot have this method run in parameter scope because the first
build on friday would have `SKIP_UNTESTED_ARTIFACTS` set to True and
the next ones set to False. Jenkins only updates the default value
after the first instanciation of the job.